+++ packages/vanilla/src/components/index.js
@@ -1,8 +1,11 @@
-export * from "./ProductCard";
-export * from "./SearchBar";
-export * from "./ProductList";
 export * from "./CartItem";
 export * from "./CartModal";
-export * from "./Toast";
-export * from "./Logo";
+export * from "./ErrorContent";
 export * from "./Footer";
+export * from "./LoadingContent";
+export * from "./Logo";
+export * from "./ProductCard";
+export * from "./ProductDetail";
+export * from "./ProductList";
+export * from "./SearchBar";
+export * from "./Toast";
